Fix mypy error caused by new numpy typing

+++ evo/tools/plot.py
@@ -596,7 +596,7 @@ def traj_xyz(axarr: np.ndarray, traj: trajectory.PosePath3D, style: str = '-',
             x = traj.timestamps
         xlabel = "$t$ (s)"
     else:
-        x = np.arange(0., len(traj.positions_xyz))
+        x = np.arange(0., len(traj.positions_xyz), dtype=float)
         xlabel = "index"
     ylabels = [
         f"$x$ ({length_unit.value})", f"$y$ ({length_unit.value})",
@@ -640,7 +640,7 @@ def traj_rpy(axarr: np.ndarray, traj: trajectory.PosePath3D, style: str = '-',
             x = traj.timestamps
         xlabel = "$t$ (s)"
     else:
-        x = np.arange(0., len(angles))
+        x = np.arange(0., len(angles), dtype=float)
         xlabel = "index"
     ylabels = ["$roll$ (deg)", "$pitch$ (deg)", "$yaw$ (deg)"]
     for i in range(0, 3):
